Fix ToggleCampus Button .setText bug
There was a minor error when someone tried to refractor the code. It is actually the other way around:         campusTextView.setText(showSGW ? "LOY" : "SGW");
If showSGW is true, it means we are on the SGW campus, therefore we want the text button to show 'LOY' so the user when he wants can toggle the button and go to LOY campus.